+++ packages/assets-controllers/src/token-prices-service/codefi-v2.ts
@@ -290,6 +290,7 @@ export const SPOT_PRICES_SUPPORT_INFO = {
   '0x1079': 'eip155:4217/slip44:60', // Tempo Mainnet - No native asset
   '0x10e6': 'eip155:4326/erc20:0x0000000000000000000000000000000000000000', // MegaETH Mainnet - Native symbol: ETH
   '0x1388': 'eip155:5000/erc20:0xdeaddeaddeaddeaddeaddeaddeaddeaddead0000', // Mantle - Native symbol: MNT
+  '0x13a7': 'eip155:5031/slip44:111115031', // Somnia Mainnet - Native symbol: SOMI
   '0x13b2': 'eip155:5042/slip44:5042', // Arc - Native symbol: USDC
   '0x1b58': 'eip155:7000/slip44:7000', // ZetaChain - Native symbol: ZETA
   '0x2105': 'eip155:8453/slip44:60', // Base - Native symbol: ETH